refactor(renderer): split oversized TabBar module
