+++ src/renderer/src/components/tab-bar/use-tab-bar-create-menu-controller.ts
@@ -1,0 +1,289 @@
+import { useEffect, useMemo, useRef, useState } from 'react'
+import { toast } from 'sonner'
+import type { TuiAgent } from '../../../../shared/types'
+import { translate } from '@/i18n/i18n'
+import { focusTerminalTabSurface } from '@/lib/focus-terminal-tab-surface'
+import { launchAgentInNewTab } from '@/lib/launch-agent-in-new-tab'
+import type { WindowsTerminalCapabilities } from '@/lib/windows-terminal-capabilities'
+import { useAppStore } from '../../store'
+import type { TabAgentLaunchOption } from './tab-agent-launch-options'
+import { buildTabCreateMenuOptions, type TabCreateMenuOption } from './tab-create-menu-options'
+import { resolveWindowsShellLaunchTarget } from './windows-shell-launch'
+import {
+  buildWindowsShellMenuEntries,
+  type WindowsShellMenuEntry
+} from './tab-bar-windows-shell-options'
+import type {
+  getProjectRuntimeShellMenuMode,
+  resolveWindowsPowerShellImplementationSetting
+} from './use-tab-bar-runtime-model'
+
+const NEW_TAB_MENU_TERMINAL_FOCUS_RETRY_MS = 50
+const NEW_TAB_MENU_TERMINAL_FOCUS_TIMEOUT_MS = 5000
+
+export type TabBarCreateMenuController = {
+  newTabMenuOpen: boolean
+  setNewTabMenuOpen: (open: boolean) => void
+  setCreateMenuQuery: (query: string) => void
+  createMenuOptions: TabCreateMenuOption[]
+  windowsShellEntries: WindowsShellMenuEntry[] | undefined
+  handleSelectCreateMenuOption: (option: TabCreateMenuOption) => void
+  launchAgentFromNewTabEntry: (agent: TuiAgent) => void
+  runPendingNewTabMenuFocusAfterClose: () => void
+  clearPendingNewTabMenuFocusOnUnmount: (node: HTMLDivElement | null) => void
+  queueNewActiveTerminalFocusAfterNewTabMenuClose: () => void
+  queueTerminalTabFocusAfterNewTabMenuClose: (tabId: string) => void
+  queueFocusAfterNewTabMenuClose: (focus: () => void) => void
+  showStaticCreateMenuItems: boolean
+}
+
+export function useTabBarCreateMenuController({
+  worktreeId,
+  resolvedGroupId,
+  terminalOnly,
+  mobileEmulatorEnabled,
+  workspaceHasSimulatorTab,
+  showWindowsShellMenu,
+  projectRuntimeShellMenuMode,
+  defaultWindowsShell,
+  defaultWindowsPowerShellImplementation,
+  windowsTerminalCapabilities,
+  agentLaunchOptions,
+  onNewTerminalTab,
+  onNewTerminalWithShell,
+  onNewBrowserTab,
+  onNewSimulatorTab,
+  onNewFileTab,
+  onOpenFileTab
+}: {
+  worktreeId: string
+  resolvedGroupId: string
+  terminalOnly: boolean
+  mobileEmulatorEnabled: boolean
+  workspaceHasSimulatorTab: boolean
+  showWindowsShellMenu: boolean
+  projectRuntimeShellMenuMode: ReturnType<typeof getProjectRuntimeShellMenuMode>
+  defaultWindowsShell: string
+  defaultWindowsPowerShellImplementation: ReturnType<
+    typeof resolveWindowsPowerShellImplementationSetting
+  >
+  windowsTerminalCapabilities: WindowsTerminalCapabilities
+  agentLaunchOptions: TabAgentLaunchOption[]
+  onNewTerminalTab: () => void
+  onNewTerminalWithShell?: (shell: string) => void
+  onNewBrowserTab: () => void
+  onNewSimulatorTab?: () => void
+  onNewFileTab?: () => void
+  onOpenFileTab?: () => void
+}): TabBarCreateMenuController {
+  // Why: <webview> clicks are out-of-process, so Radix's document-pointerdown outside-click check misses them; use window blur.
+  const [newTabMenuOpen, setNewTabMenuOpen] = useState(false)
+  const [createMenuQuery, setCreateMenuQuery] = useState('')
+  const pendingNewTabMenuFocusRef = useRef<(() => void) | null>(null)
+  const pendingNewTabMenuFocusAnimationRef = useRef<number | null>(null)
+  const pendingNewTabMenuFocusRetryRef = useRef<number | null>(null)
+  const clearPendingNewTabMenuFocusAnimation = (): void => {
+    if (pendingNewTabMenuFocusAnimationRef.current === null) {
+      return
+    }
+    cancelAnimationFrame(pendingNewTabMenuFocusAnimationRef.current)
+    pendingNewTabMenuFocusAnimationRef.current = null
+  }
+  const clearPendingNewTabMenuFocusRetry = (): void => {
+    if (pendingNewTabMenuFocusRetryRef.current === null) {
+      return
+    }
+    window.clearTimeout(pendingNewTabMenuFocusRetryRef.current)
+    pendingNewTabMenuFocusRetryRef.current = null
+  }
+  const focusNewActiveTerminalWhenReady = (
+    previousActiveTabId: string | null,
+    expiresAt: number
+  ): void => {
+    const state = useAppStore.getState()
+    if (
+      (state.activeTabType === 'terminal' || state.activeTabType === 'simulator') &&
+      state.activeTabId &&
+      state.activeTabId !== previousActiveTabId
+    ) {
+      focusTerminalTabSurface(state.activeTabId)
+      return
+    }
+    if (Date.now() >= expiresAt) {
+      return
+    }
+    pendingNewTabMenuFocusRetryRef.current = window.setTimeout(() => {
+      pendingNewTabMenuFocusRetryRef.current = null
+      focusNewActiveTerminalWhenReady(previousActiveTabId, expiresAt)
+    }, NEW_TAB_MENU_TERMINAL_FOCUS_RETRY_MS)
+  }
+  const queueNewActiveTerminalFocusAfterNewTabMenuClose = (): void => {
+    const previousActiveTabId = useAppStore.getState().activeTabId
+    pendingNewTabMenuFocusRef.current = () => {
+      // Why: paired web/SSH tab creation is async; await the host snapshot's new terminal instead of the pre-existing active tab.
+      focusNewActiveTerminalWhenReady(
+        previousActiveTabId,
+        Date.now() + NEW_TAB_MENU_TERMINAL_FOCUS_TIMEOUT_MS
+      )
+    }
+  }
+  const queueTerminalTabFocusAfterNewTabMenuClose = (tabId: string): void => {
+    pendingNewTabMenuFocusRef.current = () => focusTerminalTabSurface(tabId)
+  }
+  const queueFocusAfterNewTabMenuClose = (focus: () => void): void => {
+    pendingNewTabMenuFocusRef.current = focus
+  }
+  const windowsShellEntries = useMemo(() => {
+    return buildWindowsShellMenuEntries({
+      showWindowsShellMenu,
+      hasShellLauncher: Boolean(onNewTerminalWithShell),
+      projectRuntimeShellMenuMode,
+      defaultWindowsShell,
+      gitBashAvailable: windowsTerminalCapabilities.gitBashAvailable,
+      wslAvailable: windowsTerminalCapabilities.wslAvailable
+    })
+  }, [
+    defaultWindowsShell,
+    onNewTerminalWithShell,
+    projectRuntimeShellMenuMode,
+    showWindowsShellMenu,
+    windowsTerminalCapabilities.gitBashAvailable,
+    windowsTerminalCapabilities.wslAvailable
+  ])
+  const createMenuOptions = useMemo(
+    () =>
+      buildTabCreateMenuOptions({
+        terminalOnly,
+        windowsShellEntries,
+        hasNewBrowser: !terminalOnly,
+        hasNewMarkdown: !terminalOnly && Boolean(onNewFileTab),
+        hasOpenMarkdown: !terminalOnly && Boolean(onOpenFileTab),
+        hasSimulator: !terminalOnly && mobileEmulatorEnabled && Boolean(onNewSimulatorTab),
+        simulatorIsGoTo: workspaceHasSimulatorTab
+      }),
+    [
+      mobileEmulatorEnabled,
+      onNewFileTab,
+      onNewSimulatorTab,
+      onOpenFileTab,
+      terminalOnly,
+      windowsShellEntries,
+      workspaceHasSimulatorTab
+    ]
+  )
+  const handleSelectCreateMenuOption = (option: TabCreateMenuOption): void => {
+    switch (option.kind) {
+      case 'new-terminal':
+        queueNewActiveTerminalFocusAfterNewTabMenuClose()
+        onNewTerminalTab()
+        break
+      case 'new-terminal-shell':
+        if (!onNewTerminalWithShell || !option.shell) {
+          break
+        }
+        queueNewActiveTerminalFocusAfterNewTabMenuClose()
+        onNewTerminalWithShell(
+          resolveWindowsShellLaunchTarget(
+            option.shell,
+            defaultWindowsPowerShellImplementation,
+            windowsTerminalCapabilities.pwshAvailable
+          )
+        )
+        break
+      case 'new-browser':
+        onNewBrowserTab()
+        break
+      case 'new-markdown':
+        onNewFileTab?.()
+        break
+      case 'open-markdown':
+        onOpenFileTab?.()
+        break
+      case 'new-simulator':
+      case 'go-to-simulator':
+        onNewSimulatorTab?.()
+        break
+    }
+  }
+  const launchAgentFromNewTabEntry = (agent: TuiAgent): void => {
+    const option = agentLaunchOptions.find((candidate) => candidate.agent === agent)
+    const result = launchAgentInNewTab({
+      agent,
+      worktreeId,
+      groupId: resolvedGroupId,
+      launchSource: 'tab_bar_quick_launch'
+    })
+    if (!result) {
+      toast.error(
+        translate(
+          'auto.components.tab.bar.TabBar.ab589350e5',
+          'Could not build launch command for {{value0}}.',
+          { value0: option?.label ?? agent }
+        )
+      )
+      return
+    }
+    if (result.tabId) {
+      queueTerminalTabFocusAfterNewTabMenuClose(result.tabId)
+      return
+    }
+    queueNewActiveTerminalFocusAfterNewTabMenuClose()
+  }
+  const runPendingNewTabMenuFocusAfterClose = (): void => {
+    const pendingFocus = pendingNewTabMenuFocusRef.current
+    pendingNewTabMenuFocusRef.current = null
+    clearPendingNewTabMenuFocusAnimation()
+    clearPendingNewTabMenuFocusRetry()
+    if (pendingFocus) {
+      pendingNewTabMenuFocusAnimationRef.current = requestAnimationFrame(() => {
+        pendingNewTabMenuFocusAnimationRef.current = null
+        pendingFocus()
+      })
+    }
+  }
+  const clearPendingNewTabMenuFocusOnUnmountRef = useRef<
+    ((node: HTMLDivElement | null) => void) | null
+  >(null)
+  if (clearPendingNewTabMenuFocusOnUnmountRef.current === null) {
+    clearPendingNewTabMenuFocusOnUnmountRef.current = (node: HTMLDivElement | null): void => {
+      if (node !== null) {
+        return
+      }
+      // Why: cancel the delayed focus handoff via this root ref cleanup, avoiding an otherwise cleanup-only React Effect.
+      clearPendingNewTabMenuFocusAnimation()
+      clearPendingNewTabMenuFocusRetry()
+    }
+  }
+  const clearPendingNewTabMenuFocusOnUnmount = clearPendingNewTabMenuFocusOnUnmountRef.current
+
+  useEffect(() => {
+    if (!newTabMenuOpen) {
+      return
+    }
+    const dismiss = (): void => setNewTabMenuOpen(false)
+    window.addEventListener('blur', dismiss)
+    return () => window.removeEventListener('blur', dismiss)
+  }, [newTabMenuOpen])
+
+  useEffect(() => {
+    if (!newTabMenuOpen) {
+      setCreateMenuQuery('')
+    }
+  }, [newTabMenuOpen])
+
+  return {
+    newTabMenuOpen,
+    setNewTabMenuOpen,
+    setCreateMenuQuery,
+    createMenuOptions,
+    windowsShellEntries,
+    handleSelectCreateMenuOption,
+    launchAgentFromNewTabEntry,
+    runPendingNewTabMenuFocusAfterClose,
+    clearPendingNewTabMenuFocusOnUnmount,
+    queueNewActiveTerminalFocusAfterNewTabMenuClose,
+    queueTerminalTabFocusAfterNewTabMenuClose,
+    queueFocusAfterNewTabMenuClose,
+    showStaticCreateMenuItems: createMenuQuery.trim().length === 0
+  }
+}
